atirando com o mouse

+++ src/com/zgame/entities/BulletShoot.java
@@ -8,13 +8,13 @@ import com.zgame.main.Game;
 import com.zgame.world.Camera;
 
 public class BulletShoot extends Entity{
-	private int dx, dy;
+	private double dx, dy;
 	private double speed = 4;
 	
 	private int currentLife = 0;
 	private static final int MAX_LIFE = 10;
 
-	public BulletShoot(int x, int y, int width, int height, BufferedImage sprite, int dx, int dy) {
+	public BulletShoot(int x, int y, int width, int height, BufferedImage sprite, double dx, double dy) {
 		super(x, y, width, height, sprite);
 		this.dx = dx;
 		this.dy = dy;
